+++ tests/components/automation/test_event.py
@@ -6,6 +6,7 @@ from homeassistant.setup import setup_component
 import homeassistant.components.automation as automation
 
 from tests.common import get_test_home_assistant, mock_component
+from tests.components.automation import common
 
 
 # pylint: disable=invalid-name
@@ -50,7 +51,7 @@ class TestAutomationEvent(unittest.TestCase):
         self.assertEqual(1, len(self.calls))
         assert self.calls[0].context is context
 
-        automation.turn_off(self.hass)
+        common.turn_off(self.hass)
         self.hass.block_till_done()
 
         self.hass.bus.fire('test_event')
@@ -75,7 +76,7 @@ class TestAutomationEvent(unittest.TestCase):
         self.hass.block_till_done()
         self.assertEqual(1, len(self.calls))
 
-        automation.turn_off(self.hass)
+        common.turn_off(self.hass)
         self.hass.block_till_done()
 
         self.hass.bus.fire('test_event')
