+++ tests/components/automation/test_state.py
@@ -12,6 +12,7 @@ import homeassistant.components.automation as automation
 from tests.common import (
     fire_time_changed, get_test_home_assistant, assert_setup_component,
     mock_component)
+from tests.components.automation import common
 
 
 # pylint: disable=invalid-name
@@ -68,7 +69,7 @@ class TestAutomationState(unittest.TestCase):
             'state - test.entity - hello - world - None',
             self.calls[0].data['some'])
 
-        automation.turn_off(self.hass)
+        common.turn_off(self.hass)
         self.hass.block_till_done()
         self.hass.states.set('test.entity', 'planet')
         self.hass.block_till_done()
@@ -370,7 +371,7 @@ class TestAutomationState(unittest.TestCase):
         self.hass.states.set('test.entity_1', 'world')
         self.hass.states.set('test.entity_2', 'world')
         self.hass.block_till_done()
-        automation.turn_off(self.hass)
+        common.turn_off(self.hass)
         self.hass.block_till_done()
 
         fire_time_changed(self.hass, dt_util.utcnow() + timedelta(seconds=10))
