+++ tests/components/automation/test_time.py
@@ -11,6 +11,7 @@ import homeassistant.components.automation as automation
 from tests.common import (
     fire_time_changed, get_test_home_assistant, assert_setup_component,
     mock_component)
+from tests.components.automation import common
 
 
 # pylint: disable=invalid-name
@@ -52,7 +53,7 @@ class TestAutomationTime(unittest.TestCase):
         self.hass.block_till_done()
         self.assertEqual(1, len(self.calls))
 
-        automation.turn_off(self.hass)
+        common.turn_off(self.hass)
         self.hass.block_till_done()
 
         fire_time_changed(self.hass, dt_util.utcnow().replace(hour=0))
